change hash keys from symbols to strings, so they place nice with the YAML generated templates
the hash generated from the YML template files has Strings for keys. Because of this, our calls
to data[:some_key] were always returning nil, even if data["some_key"] existed. I first considered
symbolizing the hash generated from the YML file, but ran into issues because we're formatting
the keys, i.e. camelizing, and we can't do that if we convert the keys from strings to symbols.